[patch] moved functions to mas.py
